convert tests to use jest
