+++ packages/apollo-cache-inmemory/package.json
@@ -20,53 +20,43 @@
   },
   "homepage": "https://github.com/apollographql/apollo-client#readme",
   "scripts": {
-    "pretest": "npm run build",
-    "test": "npm run test-only --",
-    "posttest": "npm run lint",
-    "test-only": "mocha --reporter spec --full-trace dist/tests/tests.js",
-    "test-watch": "mocha --reporter spec --full-trace dist/tests/tests.js --watch",
-    "coverage": "istanbul cover ./node_modules/.bin/_mocha -- --reporter dot --full-trace dist/tests/tests.js",
-    "postcoverage": "remap-istanbul --input coverage/coverage.json --type lcovonly --output coverage/lcov.info",
-    "lint": "tslint --type-check -p tsconfig.json src/*.ts && tslint --type-check -p tsconfig.json tests/*.ts",
-    "prebuild": "npm run clean:dist",
+    "test": "jest",
+    "posttest": "npm run lint && npm run build",
+    "lint": "tslint --type-check -p tsconfig.json src/*.ts",
+    "prebuild": "npm run clean",
     "build": "tsc -p .",
-    "postbuild": "cp -R ./dist/src/. ./lib",
     "watch": "tsc -w -p .",
-    "clean": "npm run clean:dist && npm run clean:coverage",
-    "clean:dist": "rimraf dist/* && rimraf lib/*",
-    "clean:coverage": "rimraf coverage/*",
-    "prepublishOnly": "npm run clean && npm run build",
-    "compile:browser": "browserify ./dist/src/index.js --i apollo-cache-core --i apollo-utilities -o=./dist/index.js && npm run minify:browser",
-    "minify:browser": "uglifyjs -c -m -o ./dist/index.min.js -- ./dist/index.js",
+    "clean": "rimraf coverage/* && rimraf lib/*",
+    "prepublishOnly": "npm run build",
+    "compile:browser": "browserify ./lib/index.js --i apollo-cache-core --i apollo-utilities -o=./lib/bundle.js && npm run minify:browser",
+    "minify:browser": "uglifyjs -c -m -o ./lib/bundle.min.js -- ./lib/bundle.js",
     "filesize": "npm run compile:browser"
   },
   "dependencies": {
-    "apollo-utilities": "*",
-    "apollo-cache-core": "*",
-    "graphql-anywhere": "^4.0.0"
+    "apollo-utilities": "0.1.0",
+    "apollo-cache-core": "0.1.0",
+    "graphql-anywhere": "4.0.0"
   },
   "devDependencies": {
-    "@types/chai": "^4.0.0",
-    "@types/chai-as-promised": "0.0.31",
     "@types/graphql": "~0.9.0",
+    "@types/jest": "^20.0.7",
     "@types/lodash": "^4.14.72",
-    "@types/mocha": "^2.2.31",
-    "@types/sinon": "^2.3.2",
     "browserify": "^14.4.0",
-    "chai": "^4.0.2",
-    "chai-as-promised": "^7.0.0",
-    "fetch-mock": "^5.11.0",
     "graphql": "^0.10.3",
     "graphql-tag": "^2.4.2",
-    "istanbul": "^0.4.4",
+    "jest": "^20.0.4",
     "lodash": "^4.17.4",
-    "mocha": "^3.2.0",
-    "remap-istanbul": "^0.9.0",
     "rimraf": "^2.5.4",
-    "sinon": "^2.3.4",
-    "source-map-support": "^0.4.5",
+    "ts-jest": "^20.0.10",
     "tslint": "^5.0.0",
     "typescript": "^2.2.1",
     "uglifyjs": "^2.4.11"
+  },
+  "jest": {
+    "transform": {
+      ".(ts|tsx)": "<rootDir>/node_modules/ts-jest/preprocessor.js"
+    },
+    "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
+    "moduleFileExtensions": ["ts", "tsx", "js", "json"]
   }
 }
