--- packages/apollo-cache-inmemory/tests/tests.ts
@@ -1,4 +1,0 @@
-import './cache';
-import './diffAgainstStore';
-import './readFromStore';
-import './writeToStore';
